Rework entire script
The basic concept is kept, but instead of showing a flatlist,
an actual tree (corresponding to the file hierarchy) is built,
using the 'tree' program plus some extra processing with a perl
script.
This actually fits the 'filetree' name better.

There are more features planned, but the basics work. This includes:
- showing the tree
- opening the files selected in the tree
- highlighting the currently opened files
- fuzzy opener with `filetree-edit`

Planned features:
- options to pass to the `tree` program
- selecting all opened files
- creating files/directory in the tree view